refactor(operate): 代码重构

+++ rc-application/rc-app-operate-biz/src/main/java/com/rc/cloud/app/operate/domain/service/BrandDomainService.java
@@ -6,6 +6,7 @@ import com.rc.cloud.app.operate.domain.model.brand.BrandRepository;
 import com.rc.cloud.app.operate.domain.model.brand.valobj.BrandId;
 import com.rc.cloud.app.operate.domain.model.product.ProductRepository;
 import com.rc.cloud.common.core.util.AssertUtils;
+import com.rc.cloud.common.core.util.object.ObjectUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -35,7 +36,10 @@ public class BrandDomainService {
         if(productRepository.existsByBrandId(brandId)){
             throw new DomainException("已关联产品,删除失败");
         }
+        Brand brand = brandRepository.findById(brandId);
+        if (ObjectUtils.isNull(brand)) {
+            throw new DomainException("品牌唯一标识无效");
+        }
         return brandRepository.removeById(brandId);
     }
-
 }
